pyproject.toml
[project]
name = "stockly"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "yfinance (>=0.2.51,<0.3.0)",
    "requests (>=2.32.3,<3.0.0)",
    "markdown (>=3.7,<4.0)",
    "openai (>=1.59.7,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi[standard] (>=0.115.6,<0.116.0)",
    "pyright (>=1.1.392.post0,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "boto3 (>=1.36.19,<2.0.0)",
    "vercel (>=0.2.1,<0.3.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "stockly"}]